requirements.txt
absl_py==2.1.0
albumentations==1.4.17
Cython==3.0.11
fastapi==0.115.0
h5py==3.12.1
httpx==0.27.2
mdurl==0.1.2
mpmath==1.3.0
lmdb==1.5.1
matplotlib==3.9.2
mock==5.1.0
numpy==1.26.4
object_detection==0.0.3
onnx==1.16.2
onnxruntime==1.19.2
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==10.4.0
pip==24.2
protobuf==3.20.3
pydantic==2.9.2
pypandoc==1.13
PyYAML==6.0.2
Requests==2.32.3
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
scikit-image==0.24.0
starlette==0.38.6
tensorflow==2.17.0
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.5
uvicorn==0.31.0
pyzbar==0.1.9